[Decomp][Object] Cohesion split: extract field accessors to ObjectValues.cpp
